fix(ProLayout): 菜单审阅项集中修复 — TopNavHeader、样式注册、节点抽取
- TopNavHeader 移除对 menuProps 的重复展开。
- BaseMenu：抽离 MenuUtil / MenuItemTooltip 至 baseMenuNavNodes；图标 img 使用 alt=""
-；侧栏统一 mode=vertical；menuProps 合并顺序为先 menuProps 再组件样式。
- layoutNavCssVars：侧栏默认字号 14px、字重 400；useStyle 注册名含 prefixCls。
- ProLayoutNavMenu ref 类型与 eslint 递归调用注释；menu 样式断言 CSSInterpolation。
- ColorPicker：导出类型断言消除 tsc TS2742。
- 布局测试选择器与新快照；新增顶栏 menuProps 单测。
