+++ src/form/components/ColorPicker/index.tsx
@@ -51,4 +51,6 @@ const ProFormColorPicker: React.ForwardRefRenderFunction<
   );
 };
 
-export default React.forwardRef(ProFormColorPicker);
+export default React.forwardRef(ProFormColorPicker) as React.ForwardRefExoticComponent<
+  ProFormColorPickerProps & React.RefAttributes<any>
+>;
