+++ tests/layout/mobile.test.tsx
@@ -86,7 +86,7 @@ describe('mobile BasicLayout', () => {
     await waitFor(
       () => {
         html.baseElement
-          .querySelectorAll('ul.ant-pro-base-menu-inline-submenu-children')
+          .querySelectorAll('ul.ant-pro-base-menu-vertical-submenu-children')
           .forEach((ul) => {
           if (/ant-motion-collapse-(enter|leave)/.test(ul.className)) {
             throw new Error('menu motion');
